fix data type to float32 in reshape_frame
Former-commit-id: 85b4070bac5b75f04ff1123022ef1b39b35a18ab
Former-commit-id: 85fae77c63bfb2282f0d47aff98ba7ada137794e
Former-commit-id: 286a8314f9b3e64d0c04016369b52c9b8e411ff8 [formerly c922015af6f9b4ee1034634db22d452b70520ebd]
Former-commit-id: 66f597f346cddf9cbec3900184c3309ce7fa8672

+++ 2_Implementation/py/transient_visualizer/main.py
@@ -1,6 +1,3 @@
-# Press Shift+F10 to execute it or replace it with your code.
-# Press Double Shift to search everywhere for classes, files, tool windows, actions, and settings.
-
 import sys
 import getopt
 import os
@@ -123,10 +120,10 @@ def reshape_frame(files):
         np_pt = np.float32
 
     # Define an empty matrix of size image_height x image_width x temporal_samples for each channel
-    frame_A = np.empty([len(files), size[1], size[0]])
-    frame_R = np.empty([len(files), size[1], size[0]])
-    frame_G = np.empty([len(files), size[1], size[0]])
-    frame_B = np.empty([len(files), size[1], size[0]])
+    frame_A = np.empty([len(files), size[1], size[0]], dtype=np.float32)
+    frame_R = np.empty([len(files), size[1], size[0]], dtype=np.float32)
+    frame_G = np.empty([len(files), size[1], size[0]], dtype=np.float32)
+    frame_B = np.empty([len(files), size[1], size[0]], dtype=np.float32)
 
     # Set initial value to the global min e global max (updated later)
     global_max = 0
